lib/uklock: Make mutex SMP safe
We can safely acquire a lock on uniprocessor systems by disabling the
interrupts, but this does not work in an SMP environment. Hence, to
make the mutex SMP safe, this commit uses atomic compare exchange.
The commit also makes recursive locking optional (disabled by default).
While this is a breaking change, recursive locks are discouraged and
should only be used with caution. The current Unikraft core does not
use recursive locks anywhere.

Co-authored-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Reviewed-by: Andra Paraschiv <andra@unikraft.org>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #476