lib/ukvmem: Synchronize the virtual memory API
This commit provides synchronization of the mapping, unmapping
and page fault functions provided by ukvmem, by synchronizing
the functions they call.

The main point that requires synchronization is the list of VMAs
associated to one VAS. We synchronize it using read write locks,
so that concurrent traversals of this list are possible.

However, the mapping function needs to iterate trough the VMAs
list and then modify it atomically, so that we don't have two
`uk_vma_map()` calls mapping the same memory region. Therefore,
mapping calls are serialized at this point by using a simple spinlock.

Signed-off-by: Razvan Virtan <virtanrazvan@gmail.com>